% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidyieldfns.R
\name{covidcalcyield}
\alias{covidcalcyield}
\title{Estimate COVID-19 testing yield under a number of strategies}
\usage{
covidcalcyield(
  data,
  y,
  strategies,
  weight = "No Weights",
  cluster = "No Clusters",
  stratlabels = NULL
)
}
\arguments{
\item{data}{Input dataset}

\item{y}{indicator that COVID-19 test is positive}

\item{weight}{name of weight variable, defaults to "No Weights"}

\item{cluster}{name of variable indicating clusters, defaults to "No Clusters"}

\item{stratlabels}{optional vector of strategy labels to improve readability}

\item{strategyind}{name of variable indicating whether or not a specific person would be tested under a given strategy}

\item{strata}{name of variable to be stratified by (defaults to NULL)}
}
\description{
This function summarizes results across multiple strategies
}
\examples{
covidcalcyield(data = mydata, y = "COVID-19", strategies = c("str1", "str2", "str3"), stratlabels = c("All venues", "Some other strategy", "Another one"), weight="weight", cluster = "sitespotid")
}
