% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidyieldfns.R
\name{covidplotyield}
\alias{covidplotyield}
\title{A function to plot expeceted COVID-19 testing yield by strategy}
\usage{
covidplotyield(results)
}
\arguments{
\item{results}{Matrix of results from covidcalcyield function}
}
\description{
This function plots the number needed to test to identify 1 new (previously undiagnosed) case of COVID-19 and 95\% CI for each strategy
}
\examples{
covidplotyield(allresults)
}
